refactor: broadcast all settled liabilities transactions
